pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "tvd2rdf._version.__version__"}

[tool.setuptools.package-data]
tvd2rdf = ["config.yaml"]

# pyproject.toml
[tool.pytest.ini_options]
minversion = "6.2"
testpaths = [
    "tests",
]
filterwarnings = 'ignore:Cannot convert column'

[project]
name = "tvd2rdf"
dynamic = ["version"]
description = "Creates an RDF/S (+just enough OWL) vocabulary definition file based on a CSV file."
readme = "README.md"
authors = [{name = "Phil Barker", email="phil.barker@pjjk.co.uk"}]

classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Ontologists",
    "Intended Audience :: Information Technology",
    "Natural Language :: English",
]
keywords = ["RDF", "metadata application profile", "SHACL"]
dependencies = [
    "rdflib >=7.0.0",
    "otsrdflib >= 0.5",
    "PyYAML >=6.0.0", 
]
requires-python = ">=3.12"

[project.optional-dependencies]
dev = [
    "pytest >= 6.2",
    "black >= 21.6b0",
    "pytest-black >= 0.3.12",
    "pyinstaller >= 5.4"
]

[project.scripts]
tvd2rdf = "tvd2rdf.__main__:main"